requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.2
cloudinary==1.44.0
colorama==0.4.6
deprecation==2.1.0
Django==5.2.1
django-cloudinary-storage==0.3.0
djangorestframework==3.16.0
frozenlist==1.6.0
gotrue==2.12.0
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
multidict==6.4.4
outcome==1.3.0.post0
packaging==25.0
pillow==11.2.1
pluggy==1.6.0
postgrest==1.0.2
propcache==0.3.1
psycopg2-binary==2.9.10
py==1.11.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
PySocks==1.7.1
pytest==8.3.5
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
realtime==2.4.3
requests==2.32.3
selenium==4.26.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.2
supafunc==0.9.4
trio==0.27.0
trio-websocket==0.11.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
websocket-client==1.8.0
websockets==14.2
whitenoise==6.9.0
wsproto==1.2.0
yarl==1.20.0
